Add the my_all method
